+++ src/apps/automation-list.ts
@@ -1,4 +1,5 @@
-import { ApplicationRenderOptions } from "foundry-pf2e/foundry/client-esm/applications/_types.js";
+import { ApplicationRenderOptions } from "foundry-pf2e/foundry/client/applications/_module.mjs";
+import { FormDataExtended } from "foundry-pf2e/foundry/client/applications/ux/_module.mjs";
 
 const { ApplicationV2, HandlebarsApplicationMixin } = foundry.applications.api;
 
@@ -57,7 +58,7 @@ export class AutomationList extends HandlebarsApplicationMixin(ApplicationV2) {
 
         if (!disabledFiles.equals(game.settings.get("pf2e-assistant", "disabledFiles"))) {
             await game.settings.set("pf2e-assistant", "disabledFiles", disabledFiles);
-            SettingsConfig.reloadConfirm({ world: true });
+            foundry.applications.settings.SettingsConfig.reloadConfirm({ world: true });
         }
     }