+++ src/data/consumables/bendy-arm-mutagen.ts
@@ -1,60 +1,60 @@
 import { Assistant } from "assistant.ts";
 
-export const path = ["Consumables", "Choker-Arm Mutagen"];
+export const path = ["Consumables", "Bendy-Arm Mutagen"];
 
 export const actions: Assistant.Action[] = [
     {
         trigger: "consumable",
-        predicate: ["item:choker-arm-mutagen-lesser"],
+        predicate: ["item:bendy-arm-mutagen-lesser"],
         process: async (data: Assistant.Data) => {
             if (!data.speaker) return;
             const target = data.target ?? data.speaker;
 
             await game.assistant.socket.addEffect(
                 target.actor,
-                PF2E_EQUIPMENT_EFFECTS["effect-choker-arm-mutagen-lesser"],
+                PF2E_EQUIPMENT_EFFECTS["effect-bendy-arm-mutagen-lesser"],
                 { origin: data.speaker, item: data.item, target: target }
             );
         }
     },
     {
         trigger: "consumable",
-        predicate: ["item:choker-arm-mutagen-moderate"],
+        predicate: ["item:bendy-arm-mutagen-moderate"],
         process: async (data: Assistant.Data) => {
             if (!data.speaker) return;
             const target = data.target ?? data.speaker;
 
             await game.assistant.socket.addEffect(
                 target.actor,
-                PF2E_EQUIPMENT_EFFECTS["effect-choker-arm-mutagen-moderate"],
+                PF2E_EQUIPMENT_EFFECTS["effect-bendy-arm-mutagen-moderate"],
                 { origin: data.speaker, item: data.item, target: target }
             );
         }
     },
     {
         trigger: "consumable",
-        predicate: ["item:choker-arm-mutagen-greater"],
+        predicate: ["item:bendy-arm-mutagen-greater"],
         process: async (data: Assistant.Data) => {
             if (!data.speaker) return;
             const target = data.target ?? data.speaker;
 
             await game.assistant.socket.addEffect(
                 target.actor,
-                PF2E_EQUIPMENT_EFFECTS["effect-choker-arm-mutagen-greater"],
+                PF2E_EQUIPMENT_EFFECTS["effect-bendy-arm-mutagen-greater"],
                 { origin: data.speaker, item: data.item, target: target }
             );
         }
     },
     {
         trigger: "consumable",
-        predicate: ["item:choker-arm-mutagen-major"],
+        predicate: ["item:bendy-arm-mutagen-major"],
         process: async (data: Assistant.Data) => {
             if (!data.speaker) return;
             const target = data.target ?? data.speaker;
 
             await game.assistant.socket.addEffect(
                 target.actor,
-                PF2E_EQUIPMENT_EFFECTS["effect-choker-arm-mutagen-major"],
+                PF2E_EQUIPMENT_EFFECTS["effect-bendy-arm-mutagen-major"],
                 { origin: data.speaker, item: data.item, target: target }
             );
         }