+++ src/data/spells/04/whispers-of-the-void.ts
@@ -14,7 +14,12 @@ export const actions: Assistant.Action[] = [
 
             reroll.updateCondition.push(
                 ...(await game.assistant.socket.addCondition(data.speaker.actor, "persistent-damage", {
-                    persistent: { formula: `${Math.floor(data.item.rank / 2)}d8`, damageType: "void", dc: 15 }
+                    persistent: {
+                        formula: `${Math.floor(data.item.rank / 2)}d8`,
+                        damageType: "void",
+                        dc: 15,
+                        criticalHit: false
+                    }
                 }))
             );
 
@@ -32,7 +37,12 @@ export const actions: Assistant.Action[] = [
 
             reroll.updateCondition.push(
                 ...(await game.assistant.socket.addCondition(data.speaker.actor, "persistent-damage", {
-                    persistent: { formula: `${Math.floor(data.item.rank / 2) * 2}d8`, damageType: "void", dc: 15 }
+                    persistent: {
+                        formula: `${Math.floor(data.item.rank / 2) * 2}d8`,
+                        damageType: "void",
+                        dc: 15,
+                        criticalHit: false
+                    }
                 }))
             );
 
@@ -54,7 +64,12 @@ export const actions: Assistant.Action[] = [
 
             reroll.updateCondition.push(
                 ...(await game.assistant.socket.addCondition(data.speaker.actor, "persistent-damage", {
-                    persistent: { formula: `${Math.floor(data.item.rank / 2) * 2}d8`, damageType: "void", dc: 15 }
+                    persistent: {
+                        formula: `${Math.floor(data.item.rank / 2) * 2}d8`,
+                        damageType: "void",
+                        dc: 15,
+                        criticalHit: false
+                    }
                 }))
             );