+++ src/utils/item.ts
@@ -1,4 +1,5 @@
 import { ActorPF2e, EffectTrait, ItemPF2e, ItemTrait } from "foundry-pf2e";
+import { ItemUUID } from "foundry-pf2e/foundry/common/documents/_module.mjs";
 
 export async function getItem(itemUuid: ItemUUID): Promise<ItemPF2e | null> {
     let item: Maybe<ItemPF2e> = await fromUuid<ItemPF2e>(itemUuid);